+++ www/apps/cloud/sidebar.mjs
@@ -41,8 +41,13 @@ export const sidebar = [
           },
           {
             type: "link",
-            title: "Navigation Styles",
-            path: "/navigation",
+            title: "Cloud Assistant",
+            path: "/assistant",
+          },
+          {
+            type: "link",
+            title: "Changelog",
+            path: "/changelog",
           },
         ],
       },
@@ -75,6 +80,11 @@ export const sidebar = [
                 type: "link",
                 title: "Rename Repository",
                 path: "/projects/rename-repo-branch",
+              },
+              {
+                type: "link",
+                title: "Log Drains",
+                path: "/projects/log-drains",
               },
             ],
           },
@@ -343,6 +353,11 @@ export const sidebar = [
             type: "link",
             title: "Logs",
             path: "/logs",
+          },
+          {
+            type: "link",
+            title: "Status Page",
+            path: "/status",
           },
           {
             type: "link",